+++ CHANGELOG.md
@@ -5,13 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+### Changed
+### Removed
+### Fixed
+
+- Fixed end of file exception caused by analysis that run in parallel by checking validity of token on client instantiation. [INCORE1-427](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-427)
+
 ## [0.5.4] - 2019-12-23
 
 ### Fixed
 
 - Update bridge damage type in MonteCarlo Analysis
 - Fixed error handling with formatting problem in pyIncore Client
-- Fixed end of file exception caused by analysis that run in parallel by checking validity of token on client instantiation. [INCORE1-427](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-427) 
+ 
 
 ## [0.5.3] - 2019-12-20
 pyIncore release for IN-CORE v1.0