+++ CHANGELOG.md
@@ -15,6 +15,8 @@ datasets with the same name [INCORE1-433](https://opensource.ncsa.illinois.edu/j
 - Fixed end of file exception caused by analysis that run in parallel by checking validity of token on client instantiation. [INCORE1-427](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-427)
 - Fixed url inconsistency in dfr3 tests
 - Fixed error in reading token file in windows os [INCORE-449](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-449)
+- FIxed bug in setting fragility_key in building damage[INCORE1-456](https://opensource.ncsa.illinois
+.edu/jira/browse/INCORE1-456)
 
 ### Added